feat(common)!: Make useDependencyTimeout generic and usable as type guard for the dependency list
The type guard makes the hook useful as a condition in if-statements, for example:
```tsx
const [position, setPosition] = useState<Position>();

// throws if no position received after 5 seconds
if (useDependencyTimeout(5000, [position])) {
  // type guarded - "position" is always defined
  return <p>Latest position: {position}</p>;
}

return <p>Waiting for incoming data</p>;
```

BREAKING CHANGE:
The useDependencyTimeout now returns `true` if all dependencies are defined and `false` if not all dependencies are defined yet.
Before the return value was inverted.