chore: Remove interactive playground again
There were some collisions with a required `alias.process: false` field in the root `package.json` that caused the `parcel-reporter-tc-cli` builds to break (as it never found `process.send()` since `process` was always "mocked" as `{}`).

We can revert this commit if we want to add the playground back in, but need to fix the other issue in that case.